+++ signify/authenticode/__init__.py
@@ -1,0 +1,3 @@
+from .structures import *
+from .signed_pe import *
+from .authroot import *